Move `is_otel_tracing_enabled` to file without other dependencies to remove cyclic dependencies.
